Move commandutils to completerutils
